feat: use amazon marketplace and openai classes as injected dependencies
We were using dependencies as global variables, but FastAPI provides a
native way to inject tem into route's handlers easily with Depends
functions.

To avoid the creation of a new object on each call, we're using the
Singleton design pattern to ensure only one instance of amazon
marketplace and openai repository exists.